+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## Unreleased
 
-## 1.0.0 (2018-02-??)
+## 1.0.0 (2018-02-24)
 
 ### Added